Rename inner `defend_data_capture` directory to `update_supply_chain_information`.
