+++ cypress/support/commands.js
@@ -264,7 +264,7 @@ Cypress.Commands.add(
 Cypress.Commands.add(
     'lastWorkingDay',
     () => {
-        cy.exec('env PYTHONPATH=defend_data_capture python defend_data_capture/scripts/get_last_working_day_this_month.py').then(result => {
+        cy.exec('env PYTHONPATH=update_supply_chain_information python update_supply_chain_information/scripts/get_last_working_day_this_month.py').then(result => {
             cy.wrap(result.stdout)
     })
     }
