+++ defend_data_capture/update_supply_chain_information/asgi.py
@@ -1,5 +1,5 @@
 """
-ASGI config for defend_data_capture project.
+ASGI config for update_supply_chain_information project.
 
 It exposes the ASGI callable as a module-level variable named ``application``.
 
@@ -11,6 +11,8 @@ import os
 
 from django.core.asgi import get_asgi_application
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "defend_data_capture.settings")
+os.environ.setdefault(
+    "DJANGO_SETTINGS_MODULE", "update_supply_chain_information.settings"
+)
 
 application = get_asgi_application()
