+++ defend_data_capture/update_supply_chain_information/local_settings.py
@@ -1,4 +1,4 @@
-from defend_data_capture.settings import *
+from update_supply_chain_information.settings import *
 
 ROOT_URLCONF = "defend_data_capture.local_urls"
 MIDDLEWARE += ["debug_toolbar.middleware.DebugToolbarMiddleware"]
