+++ defend_data_capture/update_supply_chain_information/local_urls.py
@@ -1,6 +1,6 @@
 import debug_toolbar
 
-from defend_data_capture.urls import *
+from update_supply_chain_information.urls import *
 
 urlpatterns += [
     path("__debug__/", include(debug_toolbar.urls)),
